Split tcp endpoint stack into its own module
